Refactor extension diff class
